+++ webapp/src/components/checklist_item.tsx
@@ -15,6 +15,7 @@ import {useDispatch, useSelector} from 'react-redux';
 import {components, ControlProps} from 'react-select';
 import styled, {css} from 'styled-components';
 import {DraggableProvided} from 'react-beautiful-dnd';
+import {DateTime} from 'luxon';
 
 import {handleFormattedTextClick} from 'src/browser_routing';
 import {
@@ -41,7 +42,7 @@ import MarkdownTextbox from 'src/components/markdown_textbox';
 import CommandInput from './command_input';
 import GenericModal from './widgets/generic_modal';
 import {BaseInput} from './assets/inputs';
-import DateTimeSelector, {DateTimeOption} from './datetime_selector';
+import DateTimeSelector, {DateTimeOption, optionFromMillis} from './datetime_selector';
 import {Mode} from './datetime_input';
 interface ChecklistItemDetailsProps {
     checklistItem: ChecklistItem;
@@ -441,6 +442,39 @@ export const ChecklistItemDetails = (props: ChecklistItemDetailsProps): React.Re
         }
     };
 
+    const makeSuggestedDateTimeOptions = (mode: Mode.DateTimeValue | Mode.DurationValue) => {
+        let dateTime = DateTime.now();
+        dateTime = dateTime.endOf('day');
+
+        const list: DateTimeOption[] = [];
+        list.push(
+            {
+                ...optionFromMillis(dateTime.toMillis(), mode),
+                label: formatMessage({defaultMessage: 'Today'}),
+                labelRHS: (<LabelRight>{dateTime.weekdayShort}</LabelRight>),
+            }
+        );
+
+        dateTime = dateTime.plus({days: 1});
+        list.push(
+            {
+                ...optionFromMillis(dateTime.toMillis(), mode),
+                label: formatMessage({defaultMessage: 'Tomorrow'}),
+                labelRHS: (<LabelRight>{dateTime.weekdayShort}</LabelRight>),
+            }
+        );
+
+        dateTime = dateTime.plus({days: 6});
+        list.push(
+            {
+                ...optionFromMillis(dateTime.toMillis(), mode),
+                label: formatMessage({defaultMessage: 'Next week'}),
+                labelRHS: (<LabelRight>{dateTime.toLocaleString({weekday: 'short', day: '2-digit', month: 'short'})}</LabelRight>),
+            }
+        );
+        return list;
+    };
+
     const [profileSelectorToggle, setProfileSelectorToggle] = useState(false);
     const assignee_id = props.checklistItem.assignee_id; // to make typescript happy
 
@@ -521,6 +555,7 @@ export const ChecklistItemDetails = (props: ChecklistItemDetailsProps): React.Re
                                             className={'icon-calendar-outline icon-16 btn-icon'}
                                         />
                                     }
+                                    suggestedOptions={makeSuggestedDateTimeOptions(Mode.DateTimeValue)}
                                     onSelectedChange={onDueDateChange}
                                     customControl={ControlComponentDueDate}
                                     customControlProps={{
@@ -743,3 +778,10 @@ const ChecklistItemEditModal = (props: ChecklistItemEditModalProps) => {
         </GenericModal>
     );
 };
+
+const LabelRight = styled.div`
+    font-weight: 400;
+    font-size: 12px;
+    line-height: 16px;
+    color: rgba(var(--center-channel-color-rgb), 0.56);
+`;
